commit remaining changes

+++ taptrade-cli-demo/trader/src/communication/api.rs
@@ -34,7 +34,7 @@ pub struct BondSubmissionRequest {
 #[derive(Debug, Deserialize)]
 pub struct OrderActivatedResponse {
 	pub order_id_hex: String,
-	pub bond_locked_until_timestamp: u128, // unix timestamp. Do not touch bond till then unless offer gets taken.
+	pub bond_locked_until_timestamp: u64, // unix timestamp. Do not touch bond till then unless offer gets taken.
 }
 
 #[derive(Debug, Serialize)]
